Added cache documentation + party dependency decoupling
Added a page for gh-pages to describe the cache functionality.
Also started with caching party participant figures.
And since that code is all over the place and doing 8 things at once, I started cleaning that up.
Target is to be able to instantiate party objects without messing up the user session