Fix copy and edit seat plan (#472)
"orientation" and "u18" in seat_block are `enum` fields and not an `int`. Therefore, single quotes are required and the field should simply be treated as %string% in lansuite database queries.

"u18" is a checkbox and no value is sent on form submit if it is unchecked. Before "u18" was treated as `int` and PHP converted an empty string to 0 which old versions of MySQL accepted. Now MySQL is not as forgiving and "u18" has to be a string, but an empty string is not accepted, therefore special handling is needed for u18.

Similar to Import.php and edit.php